+++ src/components/login/HeaderLogin.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import Button from '../Button';
-import Overlay from '../Overlay';
+import Overlay from '../overlay/Overlay';
 import LoginForm from '../../containers/LoginForm';
 
 class HeaderLogin extends React.Component {
